backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
httpx
minio
redis
prefect
spacy
scikit-learn
transformers
torch
prometheus-client
python-multipart
feedparser
selectolax
parsel
playwright
apscheduler
aiolimiter
tenacity
pyyaml
python-dateutil

pytest
pytest-asyncio
pytest-cov
pytest-mock
responses
freezegun
faker